requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-genai
pydantic
httpx
